backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
psycopg2-binary>=2.9.0
django-cloudinary-storage>=0.3.0
cloudinary>=1.30.0
django-cors-headers>=4.0.0 